DOD scraper can now skip reports whose PDFs have been downloaded.
Because the DOD scraper makes an additional HTTP request for each
report's landing page, re-archiving data means making lots of requests,
even if we're already going to skip re-downloading the report. If this
becomes a common use case for other scrapers, moving some of this into
a utils method would be helpful.