Fixes issues discovered in 3046897a687e7839379bcfa074274a61c8ccc8f0.
The problem is that we want to continue to pass the published_on
datetime so that last_published_on will be set for the next iteration
even though the current published_on is not in the year range.

The next report won't have a published_on in year_range either, but this
seems like the cleanest solution.