[usps] switch to new page checking method
The USPS site lost its pagination controls.
I wrote them, but until they fix it, the scraper is still erroring each run.
This fixes that, by having the USPS scraper check the current page to see
if there are no results (that it overshot it, and the previous page was
the final page). Costs an extra HTTP request, but it works without hard-
coding any page numbers.

/cc @LindsayYoung