For state scraper, switch published date logic.
We will now start using the posted date as the published date and
only fall back to the 'official' date as necesssary. After some spot
checking (and seeing some official dates in the future), it was
determined that the posted date was closer to what we are actually
aiming for. Closes #83.