Scrapers now support --archive, to go back to their earliest year.
inspector.year_range now takes options and an archive year, so any
scraper that passes a year to that function will support --archive.

I've updated each scraper to replace its top-of-file comment with
an actual variable declaration, and each call to year_range to pass
this variable in.

This argument will work in either form of running a scraper:

  ./igs --archive
  ./inspectors/usps.py --archive