Fix an warning from BeautifulSoup
BeautifulSoup will now print a warning if it is automatically picking a HTML library. This is pretty annoying, so I have modified the code to explicitly use the lxml library with BeautifulSoup. In addition, I have added a new method to the utils library for donwloading the HTML body and parsing it with BeautifulSoup, since that was a pattern that already existed in a few inspector scrapers. I just made it more widely used so we could reduce the number of places where BeautifulSoup is instantiated.

There are still a few scrapers that need to instantiate BeautifuLSoup instead of using the utils.beautifulsoup_from_url method, usually because they need some special processing or are sending a POST request to the server.